Include more metadata fields in GraphQL query
This brings the GraphQL version of rendered news articles pages a lot
closer to the regular, Content Store version.

Newly added fields and what they're for:

* `content_id`,
* `public_updated_at`,
* `primary_publishing_organisation`,
* `publishing_app`,
* `rendering_app` and
* `updated_at`
are all used for meta tags of the same name (via the meta_tags component).

* `public_updated_at`
is also used for the "dateModified" attribute of the NewsArticle
structured data (which I think comes from the machine_readable_metadata
component).

* `worldwide_organisations`
  - `base_path` and
  - `title`
enables the use of `worldwide_organisations` in the "From:" byline.

* `organisations`
  - `analytics_identifier`
* `world_locations`
  - `analytics_identifier`
* `worldwide_organisations`
  - `analytics_identifier`
are for populating the "govuk:world-locations" and
"govuk:organisations" meta tags.

* `taxons`
  - `web_url`
is used to populate the "about->sameAs" field in the NewsArticle
structured data.

* `ordered_related_items`,
* `ordered_related_items_overrides` and
* `suggested_ordered_related_items`
used for populating the "Related content" section of the
contextual_sidebar component.

* `withdrawn_notice`
is used for displaying the "This X was withdrawn" notice (using the
notice component).

---

The test fixtures have been updated to reflect the inclusion of these
additional fields. (And a few incidental changes have been made to the
fixtures, too, e.g. minor whitespace changes.)

I'm planning to add more fixtures, and using those to add more test
coverage in later commits.