GraphQL version of news_article schema example
This is a take on the "news_article" example content schema that some
Frontend specs make use of. We've made a GraphQL equivalent to allow us
to run the same tests with both responses.

This GraphQL-flavoured example only includes fields that are currently
requested by our `Graphql::EditionQuery`. That does imply a maintainance
burdern in keeping the query and the JSON fixtures in sync, but we
haven't figured out a smarter approach to that yet.

In the longer-term we might want these to live somewhere central like
the content schema examples. For now, there's no infrastructure for that
and these news articles examples are only relevant to Frontend as
a client.