Pull high resolution image URL for meta tags
This is [used by govuk_publishing_components][image_url] to populate the
`og:image` and `twitter:image` meta tags if it's present. When absent,
the standard URL of the image will be used. In order to provide
consistency between when the data comes from Content Store versus
GraphQL, we need to include the field in the GraphQL query

This adds a test covering this for both the Content Store and GraphQL
contexts. There is a similar shared example in
spec/support/meta_tags.rb, but it wasn't being used in this spec before,
and the approach here allows us to use a short shared example for both
contexts

[image_url]: https://github.com/alphagov/govuk_publishing_components/blob/a48903d9b24dc47b1434ee89200c8077bdc6bf48/lib/govuk_publishing_components/presenters/machine_readable/page.rb#L34-L36