Local variables for test setup instead of `let`s
To me, this doesn't feel like a bad idea in general. `let`s are useful
for defining variables used in example blocks, but are probably clutter
in this case.

Right now there's a pleasant consistency between the Content Store and
GraphQL examples in this spec, but I'm en route to changing how these
values are derived in the GraphQL context and I think moving the
variables to where they're used is going to help maintain the
skim-readability.