GraphQL fixtures equivalent to schema examples
We're trying to improve the test coverage of news articles backed by
data from Publishing API's GraphQL endpoint to complement the existing
coverage of news articles backed by data from Content Store.

(We're not confident in this as a long-term solution to the problem, but
the resulting tests are useful enough that we believe it's worth trying
out, living with for a while and forming opinions based on.)

Here we introduce a few new test fixtures representing responses from
Publishing API's GraphQL endpoint. The names of the fixtures and the
choices of content item in the responses themselves are based entirely
on the existing Content Schema examples that we already use in a lot of
Frontend's tests and which themselves represent responses from Content
API.

For that reason, you should find that these files are near-copies of
their respective namesake Content Schema examples, but only including
the fields currently specified in `Graphql::EditionQuery`, because
that's what these responses are intended to reflect.