Fix shared example parameters
This shared example is used in the case study and news article model
specs. In both it's invoked with 2 arguments, but it only had
1 parameter declared on the shared example definition.

I've taken the parameter names from other, similar shared examples.

I don't think in this case it's a big issue, but ignoring the second
argument meant not testing what we thought we were. As can be seen from
the change I've had to make in the news article spec (the specified
example, when it was actually used, didn't pass the test).