Enable use of GraphQL content in shared examples
Since the backend for news articles can either be Content Store or
Publishing API's GraphQL endpoint, we want the tests to reflect that.

This commit changes the shared examples that're used by the news article
model spec. Those shared examples can now either get their source data
from a Content Schemas example (the existing behaviour and still the
default) or alternatively they can use the local spec fixture of the
same name that represents the equivalent response from Publishing API's
GraphQL endpoint.

---

This was a lot of work, which is why I've only done this for the shared
examples that're used directly by the news article model spec so far. If
we like the approach, we can roll it out for more / all of this type of
shared example (and either fixup the changes here or do that work
separately, I'm not fussed).