Use correct base_path in system spec
This spec example was stubbing a Content API request for one content
item but doing so with another content item's base_path. The test
helpers are clearly quite forgiving, but the mismatch is going to be
a blocker for introducing stubs for Publishing API's GraphQL endpoint
(...because GraphQL queries care that the base_path is correct).

This new base_path is the one from the Content Schemas example
"news_article_history_mode_translated_arabic".