Improve shared example parameter consistency
This is similar to e7e324698f8865871eaba324681a5063b6d0a1b5, which
updated the emphasised organisations shared examples. In this case
a second argument was being passed to "it can have people" but not "it
has news image". In both of those cases, the used example will remain
the same

Also updated here for consistency is the positional argument (parameter)
naming

Note that the "it can have organisations" example - which is actually in
the emphasised organisations file - is not used anywhere in the codebase