Update presenter with contents list headers
This is a new approach to generating contents lists in Frontend,
and significantly reduces the complexity that exists in government-frontend.
We have not included the existing rules and restrictions; the contents
list will always be shown if there are H2 headings present.

Corporate information pages also have an additional H2 heading,
"Corporate information", if there are `corporate_information_groups`
available in the `details` hash. This is a section at the bottom of
the page which include  additional links to other corporate information
pages. This heading is added to the end of the `content_item.headers` array.

In addition, we also don't include nested headings (H3 and below).

Once we have our array of H2 headings, we convert the headers using
the ContentsOutline model and presenter into a format suitable to
then pass to the contents_list_with_body component.

The methods `corporate_information_heading` and `corporate_information_heading_id`
have been combined and updated so that they are in a format suitable
for the ContentsOutline model, and available to use in the view.

Audit trail:
- https://github.com/alphagov/government-frontend/blob/4c38f1d3d8600373fe49ab506c758ee05e4909a3/app/presenters/corporate_information_page_presenter.rb#L21-L31
- https://github.com/alphagov/government-frontend/blob/4c38f1d3d8600373fe49ab506c758ee05e4909a3/app/presenters/content_item/corporate_information_groups.rb#L50-L57
- https://github.com/alphagov/government-frontend/blob/4c38f1d3d8600373fe49ab506c758ee05e4909a3/app/presenters/content_item/contents_list.rb#L11
- https://github.com/alphagov/government-frontend/blob/4c38f1d3d8600373fe49ab506c758ee05e4909a3/test/presenters/corporate_information_page_presenter_test.rb#L42-L49
- https://github.com/alphagov/government-frontend/blob/90cf5386b165b3a342689744feece7c83449b153/test/presenters/content_item/contents_list_test.rb