[WIP] Add methods for document contact details
Todo:
- tests for respond_online_url, response_form? and attachment_url

The method `ways_to_respond?` has been renamed to `ways_to_respond_available?`
and updated to return either `true` or `false` to make it's use
clearer.

`ways_to_respond?` is only called in the view to show the "Ways to
respond" section if it is an open_call_for_evidence document that
has a link, email or postal address available. It was originally
returning the first contact detail available in the check, but we
only care on whether they are present or not.

Audit trail:
- https://github.com/alphagov/government-frontend/blob/ef31d532d50a1ae80e8831f2f6eac60f54e06ec9/app/presenters/call_for_evidence_presenter.rb#L75-L97
- https://github.com/alphagov/government-frontend/blob/ef31d532d50a1ae80e8831f2f6eac60f54e06ec9/test/presenters/call_for_evidence_presenter_test.rb#L107-L148