Refactor summary logic in view
There is some additional logic to add a </section> tag if the document
type is call_for_evidence, which isn't clear to follow what exactly
it is closing. This has now been moved into the previous control
flow structure.

In doing so, this has introduced some repetition in calling the
summary_banner component. There also is a hefty amount of logic in
the capture methods. Moving this all into a partial should tidy things
up a bit (see next commit).