Randomise the GraphQL A/B test ourselves
These changes are based on the recent re-engineering of the GraphQL A/B
test in Collections https://github.com/alphagov/collections/pull/4114.

In order to compare the use of Content Store, our usual backend data
source, against Publishing API's GraphQL endpoint, we've been serving
a small proportion of production traffic to the latter in an A/B test.

Until now we've been using the standard GOV.UK A/B testing tool. That
tool relies on cookies to keep track of the client's selected variant,
ensuring the same variant is served for the duration of any given
session. The downside of this is that, as an optional cookie, users can
(and largely do) opt-out of A/B tests. The result for us is that we set
our variants to receive 50% of traffic each, but in practice they only
received ~4%.

Since per-session variant consistency doesn't matter to our test, and
since it does matter that we can reliably control the proportion of
traffic, we're scrapping the use of cookies and just selecting the
variant locally in the app, per-request.

Note: I've attempted to explain all of that to clarify the decision and
this commit's diff, but, FYI, so far we have only enabled this test for
content types that're still served by Collections. News Articles haven't
actually joined the A/B test yet but we'll change that shortly by
increasing the `GRAPHQL_TRAFFIC_RATE`.

Other things that needed updating in the process:

* Removing the test coverage for the "Z" A/B test variant, because that
  doesn't exist in the new A/B test approach.
* Removing test coverage for ActionDispatch::Request objects without
  `headers`, because we were only checking the headers for the choice of
  A/B test variant that Fastly sets.
* Adding the graphql=false query param to existing requests for News
  Articles in the tests. Without this, the requests are subject to the
  randomness of the the new A/B test approach and the tests have become
  flaky.