Check content's schema_name before GraphQL request
Publishing API's GraphQL endpoint only supports a handful of content
types. For Frontend's purposes, only News Articles are actually
supported right now.

Content types map to content item `schema_name`s (e.g. "news_article").
Therefore we can figure out whether a content item should be available
from the GraphQL endpoint by getting hold of its `schema_name` and
comparing it against a list.

Until now, the logic we've been using to determine whether to request
a content item from either this experimental GraphQL endpoint or from
Content Store (the default data source) has followed these steps:

1. is there a corresponding local file?
2. is the `graphql` query param set to `true` or `false`?
3. is there an A/B test header, selecting a specific data source?
4. is the response from the GraphQL endpoint successful?
5. is the `schema_name` (from the GraphQL response) on the list?

So you can see that we're leaving the schema_name-based decision until
after all the other checks have passed and after the request has already
been made to the GraphQL endpoint.

We're now planning to incrementally direct a small percentage of traffic
to use the GraphQL endpoint as its data source as part of our A/B test.
But we want to control the proportion of traffic based on the supported
content types e.g. we'd like to serve 1% of News Articles this way (as
opposed to, say, 1% of all Frontend traffic, with the non-News Articles
requests filtered out).

To enable this level of control and to avoid overwhelming Publishing
API, we're moving the schema_name check to an earlier stage before the
request is made to GraphQL. The result is a slight change in the
ContentItemLoader logic and a change back to always making a request to
Content Store, even if the page is ultimately going to be rendered with
content item data from GraphQL.