feat: add check to reroute gone paths
- it's possible that controllers may receive gone content items due to the
  nature of routing in this app.
Routing in frontend is a mixture of traditional fixed routes and routes based on format constraints (ie the routes are effectively anything, so one has to load the content item, check the schema, then route based on that).
Some of the traditional routes have suffixes that don’t correspond to items in content-store. (so for instance there isn’t a content item for  /apply-foster-child-council/westminster - there’s one for /apply-foster-child-council then a route rule that it can have a local authority on the end)
We have to match gone items somehow, and they could be anything, so we have to match them by a format constraint. But if we put the format constraint after the traditional routes it wouldn’t match any of them, and if we put it before them it would aggressively try to match the whole URL in cases like the one above and then fail because there’s no content item (we could put in a fallback where it shaves off the end of the url and tries again, but… not ideal?
So the solution seems to be - put it after, but whenever we load a content item we do a quick check to see if it’s actually gone, then internally redirect it by dispatching it to the relevant controller. That allows us to serve the right response without actually involving a full browser redirect and introducing some new path.