Update CorporateInformationPagePresenter to use ContentsList
The CorporateInformationPagePresenter only has one
additional header to include in the contents list,
which is included if there are corporate information
groups available.

The ContentsList concern takes care of the rest of
the complexity, so the presenter only needs to worry
about passing any additional headers in the correct
format.