Update SpecialistDocumentPresenter to use ContentsList
We can now remove any contents list logic in the
presenter as the concern now takes care of it.

`#show_table_of_contents?` will still be used to
set whether the contents list is shown, which is
a boolean taken from ["details"]["show_table_of_contents"].
This can be passed to the ContentsList concern as
the `visible` argument.