Update ContentsList to accept additional headers
Some document types such as DocumentCollection have
additional headers that need to be included in the
contents list. They can be passed as a keyword
argument, which is by default set as an empty array.
They are always added to the end of the list after
the H2 headers.

Reference:
- https://github.com/alphagov/frontend/blob/3e0970653ce22888e3c20836430cf4c4a94330a9/app/presenters/document_collection_presenter.rb#L45-L55