Enable custom rates for multiple GraphQL schemas
The normal backend data source for content item data is Content Store's
"/api/content" endpoint. The existing 0.5 rate here controls the 50%
percent of traffic for news_article pages that should use Publishing
API's "/graphql/content" endpoint as its data source instead.

We're in the process of introducing additional schemas and we don't want
to start out at 50% with any of the new ones. (We're actually thinking
of starting at 0% each time, which will only really enable the schema
for testing.)

This shift to a map of schema_name => traffic rate enables a custom rate
per schema.