Add #response_form? to model
This method is used in the view to determine
if the response form is shown. It requires
the attachment_url and at least one of
email or postal address,

It was originally returning the first contact
details available in its check, but we only
care about whether they are present or
not.

Audit trail:
- https://github.com/alphagov/government-frontend/blob/922537bb02ee6fa507fae8578d3009ebdbdfb517/app/presenters/consultation_presenter.rb#L105-L107
- https://github.com/alphagov/government-frontend/blob/922537bb02ee6fa507fae8578d3009ebdbdfb517/test/presenters/consultation_presenter_test.rb#L136-L149