Introduce complete fallback behaviour for lead image
This behaviour did not actually get triggered (at least from whitehall) because we have the same duplicate behaviour in the publisher.

Whitehall used this fallback logic and produced a lead image based on the organisation default news image, meaning it got to this bit code with a lead image already set.

We are keen to remove that logic from whitehall, since it already exists here. Nonetheless, whitehall has some additional fallback logic, so it defaults to primary org, then to organisation if primary is not set, and eventually to worldwide organisation if it's available. The latter is only the case for WorldNewsStory.

The context in which this work is being done is that we are reimplementing some WH functionality, and reevaluating business logic. There is some UR taking place that might mean we can remove this behaviour altogether. We are also exploring what the similitudes are between the default news image and crest usage, since they seem to largely serve the same purpose of departmental "branding".