Remove unnecessary divs in view
These divs (and classes) are only used for testing
and there is no styling present for the classes so
we may as well remove this. We will be refactoring
(and consolidating) consultation and call for
evidence pages further in the near future, and rather
then using divs to help with separation we could
separate them as shared partials.