Introduce complete fallback behaviour for lead image across all types
The existing code did not actually get triggered (at least from whitehall)* because we have the same duplicate behaviour publisher-side. Whitehall used this fallback logic and produced a lead image based on the organisation default news image, meaning it got to this bit code in FE with a lead image already set.

We are keen to remove that logic from whitehall, since it already exists here. Nonetheless, the true functionality across types that use a lead image is sort of spread between WH and FE:

- NewsArticles: image defaults to primary organisation's image, then to the first organisation's image if primary is not set, and eventually to worldwide organisation's image if it's available. The latter is only the case for WorldNewsStory. These additional fallbacks where happening in WH code, and are now here. Test coverage was added to capture behaviour.
- Speeches - WH was unaware of any lead image fallback - this was only happening FE side here. There is some additional "fall back" to speaker image, but that's not in the same concern as the lead image. The current changes did not introduce anything new for speeches, though I added test coverage to make the behaviour explicit. Speeches have the placeholder behaviour.
- Case studies - WH falls back to primary organisation image. This was not in FE before, but it behaves the same as speeches so it can be here. Currently speeches to not render the placeholder image; that has been kept the same. Test coverage was added to capture behaviour.

The context in which this work is being done is that we are reimplementing some WH functionality as "config-driven" blocks, and reevaluating business logic. There will be some UR around this that might mean we can remove this behaviour altogether. We are also exploring what the similitudes are between the default news image and crest usage, since they seem to largely serve the same purpose of departmental "branding".

The code changes should not actually change any behaviour FE-side.

*It was probably being triggered for content publisher news articles.