Introduce new WaysToRespond concern
Consultations and call for evidence pages have
many identical methods so makes sense to DRY
this up by moving them to a concern.

They only have ways to respond if they are
currently in the "open" phase, and the document
type is the shema name.

We will be moving methods in subsequent commits.