DRY up #attachments_with_details
All three document types need a list of all
attachments that are not accessible and have an
email address users can contact. We can do this
by collecting all the attachments from all the
necessary details fields in the model, and then
pass them to a shared method in the Attachments
concern.