Copy locales for `on` to common location
Ran 'bundle exec i18n-tasks cp "formats.consultation.on" "common.on"`

We don't need to have separate keys for
this under different schema names as it
is a common translation that can be shared.

We will update consultation and call for
evidence pages, and remove their locale
keys in subsequent commits.