Follow one redirect for CSV parent content items
Some [CSV previews fail](https://govuk.sentry.io/issues/6215960168/?project=202225&query=is%3Aunresolved&referrer=issue-stream) with the following error:

```
NoMethodError
undefined method 'find' for nil (NoMethodError)

    @attachment_metadata = @content_item.dig("details", "attachments").find do |attachment|
```

This occurs when assets are published, and the parent page is then moved. CSV previews rely on being able to find their parent page in order to render properly.

We can fix this by following the redirect.

I'm not sure yet whether redirect chains will be limited to one, or whether multiple steps are possible. We'll start with one and see how we get on. An alternative option is to follow indefinitely, and that doesn't sound good. We could set a different limit, but one might well be sufficient, and at least that way we keep a cap on the request time.