Cache CSV previews for 30 mins
At present, these have the following cache-control headers:

```
Cache-Control max-age=0, private, must-revalidate
```

They're quite heavy on rendering time, and they're a "preview", so it should be OK to expect them to be indicative of the linked CSV and not perpetually updated.

This will set them to :

```
Cache-Control max-age=1800, public
```

which should alleviate some load.