Improve performance of CSV previews
We've discovered a few CSVs that have:
- many empty fields at the end of rows
- many empty rows after useful data

They tend to perform terribly. We've told the publishing team, but need to defend against this on the front end.

I also discovered that the code that capped the maximum number of rows did not always work, I think due to the newline matching not covering all cases.

This is a bit of a rewrite of the CSV parsing, as I wanted to take it row by row to be able to tackle the above issues (letting the CSV parser decide what a line is appears to be an improvement).

I've leant on the existing tests, in the hope that they faithfully represent the landscape, and have added my own to represent the blank field issues.

In local tests, these amendments reduced the end-user render time in the region of 75%.