Move `response_form` locale key from call_for_evidence to consultation
Ran 'bundle exec i18n-tasks mv "formats.call_for_evidence.response_form" "formats.consultation.response_form"'

This key is currently unused but will be updated
in the partial in the following commit.

We've moved it to `consultation` as this
is where all the other common keys
shared with call_for_evidence are.