Enable GraphQL queries for all content types
At the moment, it is possible to query GraphQL by appending a query
string (`graphql=true`) but only for schemas we have specifically listed
as being available.

We want to be able to test the rendering of all schemas in GraphQL, so
need to allow this query string to use GraphQL regardless of the schema
being available in the A/B test.

Therefore changing the logic, so adding a query string will always
result in GraphQL being used to render the page.