Make 'Read and respond' link text more descriptive
DAC have reported that the link text was not descriptive enough for users to determine its purpose or relationship to surrounding page content when navigating out of context.

The content community 2i have recommended using  the name of the consultation as the new link text to fix this.

Add some bottom margin to the heading to align with the previous design.