Remove truncation of contact link
It looks like this link was truncated when the page was originally built. I think it was mainly because this link used to exist in a fixed width one third column. Now that it is full width, I've spoken to our accessibility specialist and they said it's better if the full link text is shown.